+++ agents/main.py
@@ -1,104 +1,23 @@
 import asyncio
 from pydantic import BaseModel
-from agents import Agent, Runner, function_tool, trace
+from agents import Agent, Runner, trace
 from loguru import logger
 from utils import utils
 from tools.prompts import make_prompt, CONTEXT_PROMPT
+from tools.output import WellTestContext, ZonalTestMemory, WellTestInterpretation, AnomalyInsights
+from tools.tools import save_test_memory
 from dotenv import load_dotenv
+import streamlit as st
+from st_aggrid import AgGrid
+import pandas as pd
 
+st.set_page_config(layout="wide")
 
 #Load environment variables
 load_dotenv()
 
 #Prompts
 ANOMALY_DETECTOR_PROMPT, MEMORY_SAVER_PROMPT, INTERPRETATOR_PROMPT = make_prompt(threshold=0.1)
-
-#Agents output types
-class WellTestContext(BaseModel):
-    Date: str
-    WellName: str
-    WTLIQ: float
-    WTOil: float
-    WTTHP: float
-    WTWCT: float
-    Z1BHP: float
-    Z2BHP: float
-    Z3BHP: float
-    mean_bhp: float
-    log_diff_z1bhp_meanbhp: float
-    log_diff_z2bhp_meanbhp: float
-    log_diff_z3bhp_meanbhp: float
-    log_diff_oil: float
-    log_diff_liq: float
-    log_diff_thp: float
-    log_diff_wct: float
-
-class ZonalTestMemory(BaseModel):
-    Date: str
-    WellName: str
-    Anomaly: bool
-    AnomalyType: str
-    WTLIQ: float
-    WTOil: float
-    WTTHP: float
-    WTWCT: float
-    Z1Status: str
-    Z2Status: str
-    Z3Status: str
-    Z1BHP: float
-    Z2BHP: float
-    Z3BHP: float
-
-class WellTestInterpretation(BaseModel):
-    Date: str
-    WellName: str
-    ZonalConfiguration: str
-    Interpretation: str
-    EngineerAction: str
-    InsightsSummary: str
-    """A short 2-3 sentence of your findings after analysing the data looking for anomalies"""
-
-class AnomalyInsights(BaseModel):
-    short_summary: str
-    """A short 2-3 sentence of your findings after analysing the data looking for anomalies"""
-
-
-# Tools
-@function_tool
-def save_test_memory(welldata: WellTestContext, file_path: str, 
-                    Anomaly: bool,
-                    AnomalyType: str,
-                    Z1Status: str,
-                    Z2Status: str,
-                    Z3Status: str) -> None:
-    """
-        Save the well test data record as a comma separated line in a text file.
-        Args:
-            Welldata: The well test data record to save.
-            file_path: The path to the file where the data will be saved.
-    """
-
-    file_path =  'agents/mem.txt'
-    logger.info(f"Saving well test data in memory")
-    fields = [
-        welldata.Date,
-        welldata.WellName,
-        Anomaly,
-        AnomalyType,
-        welldata.WTLIQ,
-        welldata.WTOil,
-        welldata.WTTHP,
-        welldata.WTWCT,
-        Z1Status,
-        Z2Status,
-        Z3Status,
-        welldata.Z1BHP,
-        welldata.Z2BHP,
-        welldata.Z3BHP        
-    ]
-    
-    with open(file_path, 'a') as f:
-        f.write(', '.join(map(str, fields)) + '\n')
 
 
 # Agent team
@@ -124,13 +43,41 @@ INTERPRETATOR_AGENT = Agent(
 
 #Function to run the agents
 async def main():
+
+    #Sidebar
+    with st.sidebar:
+        st.title("WellWatch")
+        st.subheader("Oil Well Test Monitoring with Agentic AI")
+        st.write("This is a demo of Colab work on Agentic AI for oil well test monitoring.")
+        st.write("The platform uses a combination of statistics and RE expertise to enable the AI agents to detect anomalies in well test data and provide insights.")
+
     logger.info("Starting to load data")
     df = utils.load_transform_welltest_data('data/RMO_Agentic AI_train_test.xlsx')
     df = df[['Date', 'WellName', 'WTLIQ', 'WTOil', 'WTTHP', 'WTWCT', 'Z1BHP',
             'Z2BHP', 'Z3BHP', 'mean_bhp', 'log_diff_z1bhp_meanbhp',
             'log_diff_z2bhp_meanbhp', 'log_diff_z3bhp_meanbhp', 'log_diff_oil',
             'log_diff_liq', 'log_diff_thp', 'log_diff_wct']].copy()
+    
     df['Date'] = df['Date'].astype(str)
+
+    #Dates list
+    dates = df['Date'].unique().tolist()
+    dates.sort()
+    #Select date
+    selected_date = st.sidebar.selectbox("Select a well test date", dates)
+    
+    if selected_date:
+        #select the data up until the selected date
+        logger.info(f"Filtering data for date {selected_date}")
+        # Filter the dataframe to include only rows with the selected dateand all rows before it
+        df = df.loc[df['Date'] <= selected_date]
+        logger.info(f"Done filtering data for date {selected_date}")
+        df_aggrid = df.copy().round(0)
+        df_aggrid['Date'] = pd.to_datetime(df_aggrid['Date'])
+        df_aggrid = df_aggrid[['Date', 'WellName', 'WTLIQ', 'WTOil', 'WTTHP', 'WTWCT', 'Z1BHP',
+            'Z2BHP', 'Z3BHP']]
+        grid_return = AgGrid(df_aggrid, editable=True, width = 800, height=300, fit_columns_on_grid_load=True)
+
     
     df = df.iloc[13:]
 
@@ -146,19 +93,22 @@ async def main():
     logger.info(f"well test input {well_test_input}")
 
     # Ensure the entire workflow is a single trace
-    with trace("Deterministic story flow"):
+    button = st.button("Run well test AI Agents workflow")
+    if button:
+        # Run the agents in a deterministic story flow
+        with trace("Deterministic story flow"):
 
-        # Run the anomaly detection agent
-        result_anomaly = await Runner.run(anomaly_detection_agent, input=f' Here are the well test data {well_test_input.model_dump()}' , context=context)
-        print(result_anomaly.final_output)
+            # Run the anomaly detection agent
+            result_anomaly = await Runner.run(anomaly_detection_agent, input=f' Here are the well test data {well_test_input.model_dump()}' , context=context)
+            print(result_anomaly.final_output)
 
-        # Run the memory saver agent
-        result_memory = await Runner.run(MEMORY_SAVER_AGENT, input=f' Here are the well test data {well_test_input.model_dump()} and this is what the anomaly analysis result is {result_anomaly.final_output}' , context=context)
-        print(result_memory.final_output)
+            # Run the memory saver agent
+            result_memory = await Runner.run(MEMORY_SAVER_AGENT, input=f' Here are the well test data {well_test_input.model_dump()} and this is what the anomaly analysis result is {result_anomaly.final_output}' , context=context)
+            print(result_memory.final_output)
 
-        # Run the interpretator agent
-        result_interpretator = await Runner.run(INTERPRETATOR_AGENT, input=f' Here are the well test data {well_test_input.model_dump()} and this is what the anomaly analysis result is {result_anomaly.final_output}' , context=context)
-        print(result_memory.final_output)
+            # Run the interpretator agent
+            result_interpretator = await Runner.run(INTERPRETATOR_AGENT, input=f' Here are the well test data {well_test_input.model_dump()} and this is what the anomaly analysis result is {result_anomaly.final_output}' , context=context)
+            print(result_interpretator.final_output)
 
 if __name__ == "__main__":
     asyncio.run(main())
